+++ event-sourcing/subscription-manager/src/test/java/uk/gov/justice/services/event/sourcing/subscription/startup/EventCatchupProcessorTest.java
@@ -4,11 +4,11 @@ import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import uk.gov.justice.services.event.source.subscriptions.repository.jdbc.SubscriptionsRepository;
 import uk.gov.justice.services.event.sourcing.subscription.manager.EventSourceProvider;
 import uk.gov.justice.services.event.sourcing.subscription.startup.manager.EventStreamConsumerManager;
 import uk.gov.justice.services.eventsourcing.source.core.EventSource;
 import uk.gov.justice.services.messaging.JsonEnvelope;
+import uk.gov.justice.services.subscription.ProcessedEventTrackingService;
 import uk.gov.justice.subscription.domain.subscriptiondescriptor.Subscription;
 
 import java.util.stream.Stream;
@@ -26,7 +26,7 @@ import org.slf4j.Logger;
 public class EventCatchupProcessorTest {
 
     @Mock
-    private SubscriptionsRepository subscriptionsRepository;
+    private ProcessedEventTrackingService processedEventTrackingService;
 
     @Mock
     private EventSourceProvider eventSourceProvider;
@@ -55,7 +55,7 @@ public class EventCatchupProcessorTest {
 
         when(subscription.getEventSourceName()).thenReturn(eventSourceName);
         when(eventSourceProvider.getEventSource(eventSourceName)).thenReturn(eventSource);
-        when(subscriptionsRepository.getOrInitialiseCurrentEventNumber(eventSourceName)).thenReturn(eventNumber);
+        when(processedEventTrackingService.getLatestProcessedEventNumber(eventSourceName)).thenReturn(eventNumber);
         when(eventSource.findEventsSince(eventNumber)).thenReturn(Stream.of(event_1, event_2, event_3));
 
         eventCatchupProcessor.performEventCatchup(subscription);