+++ subscription-event-tracking/src/main/java/uk/gov/justice/services/subscription/ProcessedEventTrackItem.java
@@ -1,0 +1,52 @@
+package uk.gov.justice.services.subscription;
+
+import java.util.Objects;
+
+public class ProcessedEventTrackItem {
+
+    private final long previousEventNumber;
+    private final long eventNumber;
+    private final String source;
+
+    public ProcessedEventTrackItem(final long previousEventNumber, final long eventNumber, final String source) {
+        this.previousEventNumber = previousEventNumber;
+        this.eventNumber = eventNumber;
+        this.source = source;
+    }
+
+    public long getPreviousEventNumber() {
+        return previousEventNumber;
+    }
+
+    public long getEventNumber() {
+        return eventNumber;
+    }
+
+    public String getSource() {
+        return source;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (!(o instanceof ProcessedEventTrackItem)) return false;
+        final ProcessedEventTrackItem that = (ProcessedEventTrackItem) o;
+        return previousEventNumber == that.previousEventNumber &&
+                eventNumber == that.eventNumber &&
+                Objects.equals(source, that.source);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(previousEventNumber, eventNumber, source);
+    }
+
+    @Override
+    public String toString() {
+        return "ProcessedEventTrackItem{" +
+                "previousEventNumber=" + previousEventNumber +
+                ", eventNumber=" + eventNumber +
+                ", source='" + source + '\'' +
+                '}';
+    }
+}